research-workflow-system/pyproject.toml
[project]
name = "research-workflow-system"
version = "0.1.0"
description = "Multi-topic research system with memory and cross-topic learning"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langmem>=0.1.0",
    "langchain-postgres>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "tavily-python>=0.5.0",
    "psycopg2-binary>=2.9.0",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
